+++ hash/example_test.go
@@ -1,4 +1,4 @@
-package crypto
+package hash
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
@@ -7,20 +7,28 @@ package crypto
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// JCHash generate Google's Jump Consistent Hash
-// More info: http://arxiv.org/abs/1406.2294
-func JCHash(key uint64, buckets int) int {
-	var b, j int64
+import (
+	"fmt"
+)
 
-	if buckets <= 0 {
-		buckets = 1
-	}
+// ////////////////////////////////////////////////////////////////////////////////// //
 
-	for j < int64(buckets) {
-		b = j
-		key = key*2862933555777941757 + 1
-		j = int64(float64(b+1) * (float64(int64(1)<<31) / float64((key>>33)+1)))
-	}
+func ExampleJCHash() {
+	var (
+		key     uint64 = 0xDEADBEEF
+		buckets int    = 128 * 1024
+	)
 
-	return int(b)
+	hash := JCHash(key, buckets)
+
+	fmt.Println(hash)
+
+	// Output:
+	// 110191
+}
+
+func ExampleFileHash() {
+	hash := FileHash("/path/to/some/file")
+
+	fmt.Println(hash)
 }